Issue 892: move flattenLessons to course script
