+++ aemedge/templates/course/course.js
@@ -1,8 +1,7 @@
-import { getCourseData, createCourseBaseTemplate } from '../../scripts/course/course.js';
+import { getCourseData, createCourseBaseTemplate, flattenLessons } from '../../scripts/course/course.js';
 import { createElement, i18n, preserveHideParameters } from '../../scripts/utils.js';
 import { courseDataChange } from '../../scripts/actions/course.js';
 import { addCourseCertificate } from '../../scripts/course/certificate.js';
-import { flattenLessons } from '../lesson/lesson.js';
 
 async function addBeginCourseButton(courseData) {
   const main = document.querySelector('main');
